Removed SetWindowTitle()'s side-effect when enabled=false
Changed SetWindowTitle() to no longer modify the window title of the Guild Wars window when enabled=false. Previously it would force the window title to "Guild Wars" which would override any modification made by other programs.

Recommending that this function do nothing when the Toolbox option is disabled. This is the expected behavior.