Added Tooltip and Tweaked Alerts Support
Added tooltip to the button explaining functionality.

Tweaked the alerts section to handle both Kamadan and Pre-Searing. I believe this is what was referenced in the original comment, but if I missed something, feel free to let me know. Thanks, and good catch by the way!

Original comment from 3vcloud:
"This is good, but needs a hover tooltip on the button to explain what it does, and refactor in the alerts section (this would work with Ascalon messages, but not when you're in Kamadan AE1 for example)"