added two faq that have been asked in the discord
I have added a solution to the two issues; icons not loading and Font.ttf being missing. These are the solutions that, according to users that have had their issues rectified, are working. I even tried it myself on my installation. I am not sure if either of these are solved by using the solution provided in "The icons on the Pcons, Materials, and Toolbox window aren't showing up. How do I get them back?", If they are, then ignore this by all means 👍.