Fix colour of NPC allies on minimap
It would appear that the colour used for "Non-attackable allies" is always
`color_ally`, corresponding to the "Ally (player)" setting. There are
however two types of agents that would fall under this category: NPCs in
outposts which are henchmen (which should use `color_ally_npc`), and regular
allies.

The issue can be seen by going into an outpost with henchmen and observing
that the colour used is the one intended for player allies.

The fix is to check whether the agent is an NPC before deciding which of
two colours to use.

Attached are a couple of images to demonstrate.

energy#4904