Fix file paths in some CMake files
Some file paths in a couple CMake files had the wrong case.

This has been gotten away with when compiling directly from Windows, as paths are *usually* case insensitive and so, for instance, "includes/" matches "INCLUDES/".

However, on a very specific setup (using `msvc-wine` to cross-compile w/ MSVC from Linux), the host OS's CMake is used and thus the CMake files containing paths with a differing case from their names on-disk cause an issue.

By the way, it's the easiest way to cross-compile GWToolbox that I found, in case other people were interested in trying too.