Fix travel window dropdowns
f4b0b29 removed keyboard input functionality for the dropdowns, and replaced
the dropdown preview text with a label, pushing the "Go" button outside the
bounds of the window.

This reverts those changes to the dropdowns to again use MyCombo instead,
but implements the getter function in a more compact way.