Implement active quest objectives widget
Creates ActiveQuestWidget, which displays the name and objectives of the
current active quest.

This ended up way more complicated than I wanted.  I originally tried to write
it in a very simple style with an update loop checking for state changes,
rather than using packet callbacks, but then it became apparent that I needed
the callbacks to detect objectives being completed, etc.  So this probably
could do with a refactor already.  The Update call could potentially be
replaced with a function only triggered by the relevant packets, as well
as an initial call on initialize.

Another refactor reason is the string decoding.  I needed custom callbacks,
so was unable to use EncString for the objectives.