Improve RVA detection in injector exe
Sometimes Windows doesn't clean up a previous Gw.exe process fully.
When this happens, it's possible for the pid of the dead process
to be lower, and for the RVA scans to fail for that process.

This patch loops over all candidate processes until it finds one
where both character and email RVAs succeed, only failing if
no processes have both scans succeed