Improve /armory functionality
Adds costumes, weapons and unequipping to /armory.  As a consequence,
the search is now stricter and must match the name exactly rather than
just the beginning of it.

Unequips costumes when you (un)equip any conflicting armor piece

Removes Aegis of Terror from shield list as it is a duplicate skin