Add Option to toggle between different LiveSplit formats
LiveSplit One used a json command format while LiveSplit (the desktop application) wants just the raw command, not even a newline. While it is possible to convert the formats using jq on linux, the conversion is a pain on windows. With the new options the user can just use websocat to connect the two webservers.